requirements.txt
torch>=2.3.0
transformers>=4.40.0
datasets>=2.19.0
tokenizers>=0.19.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gradio>=4.31.0
groq>=0.9.0
spacy>=3.7.0
nltk>=3.8.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
mlflow>=2.13.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
pytest>=8.0.0
httpx>=0.27.0
ruff>=0.4.0
black>=24.4.0
pre-commit>=3.7.0
jupyter>=1.0.0
tqdm>=4.66.0
